Match View Controller - UI implemented
Did a quick UI to display a list of matches when the search function is
used; currently five display at a time, and I’m thinking the “more
matches” option pulls up the next several in the list.

I’m thinking that, ideally, we’d want to use infinite scrolling for an
actual to-market app, but we can cross that bridge when we come to it.

Attempted to replace the “Matches” link on the tab bar so it directs
the user to the match list (instead of an individual match), but
couldn’t quite figure that out.